requirements.txt
Django==5.0.2
python-pptx==0.6.23
PyPDF2==3.0.1
google-generativeai==0.3.2
scikit-learn==1.4.0
python-dotenv==1.0.1
Pillow==10.2.0
huggingface_hub==0.20.3

djangorestframework==3.15.1
textract==1.6.5
django-allauth==0.61.1 